fix(query): V24 save row count with int16_t smaller

+++ src/common/inc/texpr.h
@@ -72,7 +72,7 @@ struct SSchema;
 typedef struct {
   int16_t type;
   int16_t bytes;
-  int16_t numOfRows;
+  int32_t numOfRows;
   char* data;
 } tExprOperandInfo;
